Move write options validation test to correct class. Update classname and driver check to use Strings.isNullOrEmpty
